+++ src/main/java/com/recipe/app/src/fridgeBasket/application/FridgeBasketService.java
@@ -4,9 +4,9 @@ import com.recipe.app.common.exception.BaseException;
 import com.recipe.app.src.fridgeBasket.application.dto.FridgeBasketDto;
 import com.recipe.app.src.fridgeBasket.domain.FridgeBasket;
 import com.recipe.app.src.fridgeBasket.mapper.FridgeBasketRepository;
-import com.recipe.app.src.ingredient.mapper.IngredientRepository;
+import com.recipe.app.src.ingredient.infra.IngredientRepository;
 import com.recipe.app.src.ingredient.domain.Ingredient;
-import com.recipe.app.src.ingredient.mapper.IngredientCategoryRepository;
+import com.recipe.app.src.ingredient.infra.IngredientCategoryRepository;
 import com.recipe.app.src.ingredient.domain.IngredientCategory;
 import com.recipe.app.src.user.domain.User;
 import lombok.RequiredArgsConstructor;
